Move 'my uploads' to the separate component
